backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
requests==2.32.3
pillow==10.4.0
numpy==2.1.1
torch>=2.2.0
# Best-effort ViT backbone compatible with PanDerm checkpoint
# Will load with strict=False if shapes/keys differ
# You can replace with your exact PanDerm code if available
# pytorch-lightning is optional; included for broader checkpoint compat
pytorch-lightning>=2.0.0
# Optional vision backbones
# If timm fails to install on your system, comment it out and the server will still run
# with basic features and placeholders.
timm>=1.0.7

